requirements.txt
matplotlib==3.5.1
numpy==1.21.5
Pillow==9.0.1
Pillow==9.5.0
pydensecrf==1.0rc3
scikit_learn==1.0.2
scipy==1.7.3
torch==2.0.1
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.64.0
wget==3.2
